fix: keep Node-only URL safety checks out of the client bundle
The SSRF guard added node:net and node:dns/promises imports to
lib/web-search-catalog.ts, which is imported by client components;
webpack cannot bundle node: URIs, breaking the production build.
Move isPublicHttpUrl into the server-only lib/public-http-url.ts and
update the route and tests to import it from there.
